Asset Management
Created an easy way for screens to share similar assets without having
to reload them. I have no tested it yet so there might be some problems
with it. I also decided to not add the ability to create bitmap fonts at
runtime and instead are going to make them before runtime with a new
program. I hope this will save on loading times.