GUI Update 1
Added several images for button textures as well as fonts and atlases
that go with it. There is currently a bug where the MainMenu Screen does
not re-receive the setInputProcessor command. The button table for the
main menu is also not finished and draws the table centered in the
bottom right corner. The first buttons added might also be too small.